requirements.txt
# GhostTrack Requirements
# Core ML frameworks
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0
plotly>=5.14.0

# Optimization
wandb>=0.15.0
optuna>=3.2.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
